feat: Refactor FEN generation for Pawn Race and King Escape puzzles
