+++ modules/pawn-race.js
@@ -3,7 +3,7 @@ import { BoardManager } from "./BoardManager.js";
 import { ThemeManager } from "./ThemeManager.js";
 import { MovesTableController } from "./MovesTableController.js";
 import { PawnRaceGameController } from "./PawnRaceGameController.js";
-import { FenGenerator } from "./FenGenerator.js";
+import { PawnRaceFenGenerator } from "./FenGenerator.js";
 
 
 // Initialize theme manager first
@@ -11,8 +11,8 @@ const themeManager = new ThemeManager();
 
 // Main module initialization
 const chess = new Chess();
-const fenGenerator = new FenGenerator();
-const fen = fenGenerator.generatePawnRaceFen();
+const fenGenerator = new PawnRaceFenGenerator();
+const fen = fenGenerator.generateFen();
 const boardManager = new BoardManager("pawn-race-board", {
     position: fen, // Set the starting position for the board
     assetsUrl: "cm-chessboard-master/assets/",
